Move reminder/user validator to notifier directory
